parsing data from Retrofit / fix Details

+++ app/src/main/java/com/deeper/popularmovies/utils/api/model/movieList/MovieListResult.java
@@ -21,7 +21,7 @@ public class MovieListResult {
     private Boolean video;
     @SerializedName("vote_average")
     @Expose
-    private Double voteAverage;
+    private Float voteAverage;
     @SerializedName("title")
     @Expose
     private String title;
@@ -77,11 +77,11 @@ public class MovieListResult {
         this.video = video;
     }
 
-    public Double getVoteAverage() {
+    public float getVoteAverage() {
         return voteAverage;
     }
 
-    public void setVoteAverage(Double voteAverage) {
+    public void setVoteAverage(float voteAverage) {
         this.voteAverage = voteAverage;
     }
